refactor: move set_collection function to make it generic
