+++ openedx_learning/apps/authoring/components/api.py
@@ -13,18 +13,16 @@ are stored in this app.
 from __future__ import annotations
 
 import mimetypes
-from datetime import datetime, timezone
+from datetime import datetime
 from enum import StrEnum, auto
 from logging import getLogger
 from pathlib import Path
 from uuid import UUID
 
-from django.core.exceptions import ValidationError
 from django.db.models import Q, QuerySet
 from django.db.transaction import atomic
 from django.http.response import HttpResponse, HttpResponseNotFound
 
-from ..collections.models import Collection, CollectionPublishableEntity
 from ..contents import api as contents_api
 from ..publishing import api as publishing_api
 from .models import Component, ComponentType, ComponentVersion, ComponentVersionContent
@@ -51,7 +49,6 @@ __all__ = [
     "look_up_component_version_content",
     "AssetError",
     "get_redirect_response_for_component_asset",
-    "set_collections",
 ]
 
 
@@ -605,54 +602,3 @@ def get_redirect_response_for_component_asset(
     )
 
     return HttpResponse(headers={**info_headers, **redirect_headers})
-
-
-def set_collections(
-    learning_package_id: int,
-    component: Component,
-    collection_qset: QuerySet[Collection],
-    created_by: int | None = None,
-) -> set[Collection]:
-    """
-    Set collections for a given component.
-
-    These Collections must belong to the same LearningPackage as the Component, or a ValidationError will be raised.
-
-    Modified date of all collections related to component is updated.
-
-    Returns the updated collections.
-    """
-    # Disallow adding entities outside the collection's learning package
-    invalid_collection = collection_qset.exclude(learning_package_id=learning_package_id).first()
-    if invalid_collection:
-        raise ValidationError(
-            f"Cannot add collection {invalid_collection.pk} in learning package  "
-            f"{invalid_collection.learning_package_id} to component {component} in "
-            f"learning package {learning_package_id}."
-        )
-    current_relations = CollectionPublishableEntity.objects.filter(
-        entity=component.publishable_entity
-    ).select_related('collection')
-    # Clear other collections for given component and add only new collections from collection_qset
-    removed_collections = set(
-        r.collection for r in current_relations.exclude(collection__in=collection_qset)
-    )
-    new_collections = set(collection_qset.exclude(
-        id__in=current_relations.values_list('collection', flat=True)
-    ))
-    # Use `remove` instead of `CollectionPublishableEntity.delete()` to trigger m2m_changed signal which will handle
-    # updating component index.
-    component.publishable_entity.collections.remove(*removed_collections)
-    component.publishable_entity.collections.add(
-        *new_collections,
-        through_defaults={"created_by_id": created_by},
-    )
-    # Update modified date via update to avoid triggering post_save signal for collections
-    # The signal triggers index update for each collection synchronously which will be very slow in this case.
-    # Instead trigger the index update in the caller function asynchronously.
-    affected_collection = removed_collections | new_collections
-    Collection.objects.filter(
-        id__in=[collection.id for collection in affected_collection]
-    ).update(modified=datetime.now(tz=timezone.utc))
-
-    return affected_collection
